content polish with fiance
Working on the finer layout and content. bringing in new pictures in place of tables. added some styles and played around a bunch.

attempted to add a script to add a line break in the date on teh landing banner after the comma so that it wouldn't cover my face - new font makes it unnecessary